Noticket: fix PMD warnings
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@16839 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ src/java/org/ejbca/core/ejb/keybind/impl/AuthenticationKeyBinding.java
@@ -51,7 +51,7 @@ public class AuthenticationKeyBinding extends InternalKeyBindingBase {
 
     @Override
     public float getLatestVersion() {
-        return Long.valueOf(serialVersionUID).floatValue();
+        return serialVersionUID;
     }
 
     @Override
